backend/requirements.txt
Django==4.2.7
celery==5.3.1
redis==4.6.0
telethon==1.33.0
numpy==1.23.5
pandas==1.5.3
djangorestframework==3.14.0
django-cors-headers==3.14.0
Pillow==10.0.1